Modularize API routes
